setup calendar classes and renamed hook file names
