+++ components/date-picker/src/Calendar.tsx
@@ -3,12 +3,11 @@ import { twMerge } from "tailwind-merge";
 
 const calendardefaultClasses = {
 	defaultSelectableTextStyle: "bg-white hover:bg-gray-100 ",
-	defaultDateButtonStyle:
-		"inline-block w-[calc(100% / 7)] border-none rounded-md",
-	selectedOrTodayStyles: "font-semibold border border-blue-300",
-	selectedTextStyles: "text-white font-semibold bg-blue-400",
+	defaultDateButtonStyle: "inline-block border-none rounded-md",
+	selectedOrTodayStyles: "font-semibold border border-blue-200",
+	selectedTextStyles: "text-white font-semibold bg-blue-200",
 	selectableTextStyles: "text-gray-800 ",
-	todayButNotSelectedStyles: "bg-blue-300",
+	todayButNotSelectedStyles: "bg-blue-100",
 	weekendStyles: "text-red-400 ",
 	topLeftBorderStyles: "rounded-tl-lg",
 	topRightBorderStyles: "rounded-tr-lg",
@@ -21,6 +20,7 @@ const calendardefaultClasses = {
 export interface CalendarBaseClassesProps {
 	calendarWrapperClasses?: string;
 	calenderHeaderButtonsWrapper?: string;
+	calendarWeeksClasses?: string;
 	headerButtonClasses?: string;
 	singleCalenderSectionWrapper?: string;
 	monthNameClasses?: string;
@@ -72,7 +72,7 @@ export default function Calendar({
 }: CalenderProps) {
 	if (calendars.length) {
 		return (
-			<div className={calendarBaseClasses?.calendarWrapperClasses || ""}>
+			<div className={calendarBaseClasses?.calendarWrapperClasses || "grid"}>
 				<div
 					className={`flex justify-between ${
 						calendarBaseClasses?.calenderHeaderButtonsWrapper || ""
@@ -98,7 +98,7 @@ export default function Calendar({
 				{calendars.map((calendar: any) => (
 					<div
 						key={`${calendar.month}${calendar.year}`}
-						className={`box-border inline-block ${
+						className={`box-border grid inline-block ${
 							calendarBaseClasses?.singleCalenderSectionWrapper || ""
 						}`}
 					>
@@ -111,14 +111,14 @@ export default function Calendar({
 						</div>
 
 						<div
-							className={`flex justify-between px-1 ${
+							className={`grid grid-cols-7 ${
 								calendarBaseClasses?.weekNameWrapper || ""
 							}`}
 						>
 							{weekdayNamesShort.map((weekday) => (
 								<div
 									key={`${calendar.month}${calendar.year}${weekday}`}
-									className={`inline-block text-center bg-transparent border-none w-[calc(100% / 7)] ${
+									className={`inline-block text-center bg-transparent border-none ${
 										calendarBaseClasses?.weekNameClasses || ""
 									}`}
 								>
@@ -126,91 +126,97 @@ export default function Calendar({
 								</div>
 							))}
 						</div>
+						<div
+							className={twMerge(
+								"grid grid-cols-7",
+								calendarBaseClasses?.calendarWeeksClasses,
+							)}
+						>
+							{calendar.weeks.map((week: any, weekIndex: any) =>
+								week.map((dateObj: any, dayIdx: any) => {
+									const key = `${calendar.month}${calendar.year}${weekIndex}${dayIdx}`;
+									if (!dateObj) {
+										return (
+											<div
+												key={key}
+												className="inline-block px-2 bg-transparent border-none "
+											/>
+										);
+									}
+									const {
+										date,
+										selected,
+										selectable,
+										today,
+										prevMonth,
+										nextMonth,
+									} = dateObj;
 
-						{calendar.weeks.map((week: any, weekIndex: any) =>
-							week.map((dateObj: any, dayIdx: any) => {
-								const key = `${calendar.month}${calendar.year}${weekIndex}${dayIdx}`;
-								if (!dateObj) {
+									const isCurrentMonth = !prevMonth && !nextMonth;
+
 									return (
-										<div
+										<button
 											key={key}
-											className="inline-block px-2 bg-transparent border-none w-[calc(100% / 7)]"
-										/>
+											{...getDateProps({ dateObj })}
+											className={twMerge(
+												calendardefaultClasses.defaultDateButtonStyle,
+
+												selectable &&
+													twMerge(
+														calendarBaseClasses?.selectableClasses,
+														calendardefaultClasses.defaultSelectableTextStyle,
+													),
+												(selected || today) &&
+													twMerge(
+														calendarBaseClasses?.selectedOrTodayClasses,
+														calendardefaultClasses.selectedOrTodayStyles,
+													),
+												selected &&
+													twMerge(
+														calendarBaseClasses?.selectedTextClasses,
+														calendardefaultClasses.selectedTextStyles,
+													),
+												today &&
+													!selected &&
+													twMerge(
+														calendarBaseClasses?.todayButNotSelectedClasses,
+														calendardefaultClasses.todayButNotSelectedStyles,
+													),
+												selectable &&
+													twMerge(
+														calendarBaseClasses?.selectableTextClasses,
+														calendardefaultClasses.selectableTextStyles,
+													),
+												(dayIdx == 0 || dayIdx == 6) &&
+													twMerge(
+														calendarBaseClasses?.weekendClasses,
+														calendardefaultClasses.weekendStyles,
+													),
+												weekIndex === 0 &&
+													dayIdx === 0 &&
+													calendardefaultClasses.topLeftBorderStyles,
+												weekIndex === 0 &&
+													dayIdx === 6 &&
+													calendardefaultClasses.topRightBorderStyles,
+												weekIndex === calendar.weeks.length - 1 &&
+													dayIdx === 0 &&
+													calendardefaultClasses.bottomLeftBorderStyles,
+												weekIndex === calendar.weeks.length - 1 &&
+													dayIdx === 6 &&
+													calendardefaultClasses.bottomRightBorderStyles,
+												!isCurrentMonth &&
+													twMerge(
+														calendarBaseClasses?.notCurrentMonthClasses,
+														calendardefaultClasses.notCurrentMonthStyles,
+													),
+											)}
+										>
+											{selectable ? date.getDate() : "X"}
+										</button>
 									);
-								}
-								const {
-									date,
-									selected,
-									selectable,
-									today,
-									prevMonth,
-									nextMonth,
-								} = dateObj;
-
-								const isCurrentMonth = !prevMonth && !nextMonth;
-
-								return (
-									<button
-										key={key}
-										{...getDateProps({ dateObj })}
-										className={twMerge(
-											calendardefaultClasses.defaultDateButtonStyle,
-
-											selectable &&
-												twMerge(
-													calendarBaseClasses?.selectableClasses,
-													calendardefaultClasses.defaultSelectableTextStyle,
-												),
-											(selected || today) &&
-												twMerge(
-													calendarBaseClasses?.selectedOrTodayClasses,
-													calendardefaultClasses.selectedOrTodayStyles,
-												),
-											selected &&
-												twMerge(
-													calendarBaseClasses?.selectedTextClasses,
-													calendardefaultClasses.selectedTextStyles,
-												),
-											today &&
-												!selected &&
-												twMerge(
-													calendarBaseClasses?.todayButNotSelectedClasses,
-													calendardefaultClasses.todayButNotSelectedStyles,
-												),
-											selectable &&
-												twMerge(
-													calendarBaseClasses?.selectableTextClasses,
-													calendardefaultClasses.selectableTextStyles,
-												),
-											(dayIdx == 0 || dayIdx == 6) &&
-												twMerge(
-													calendarBaseClasses?.weekendClasses,
-													calendardefaultClasses.weekendStyles,
-												),
-											weekIndex === 0 &&
-												dayIdx === 0 &&
-												calendardefaultClasses.topLeftBorderStyles,
-											weekIndex === 0 &&
-												dayIdx === 6 &&
-												calendardefaultClasses.topRightBorderStyles,
-											weekIndex === calendar.weeks.length - 1 &&
-												dayIdx === 0 &&
-												calendardefaultClasses.bottomLeftBorderStyles,
-											weekIndex === calendar.weeks.length - 1 &&
-												dayIdx === 6 &&
-												calendardefaultClasses.bottomRightBorderStyles,
-											!isCurrentMonth &&
-												twMerge(
-													calendarBaseClasses?.notCurrentMonthClasses,
-													calendardefaultClasses.notCurrentMonthStyles,
-												),
-										)}
-									>
-										{selectable ? date.getDate() : "X"}
-									</button>
-								);
-							}),
-						)}
+								}),
+							)}
+						</div>
 					</div>
 				))}
 			</div>
