+++ components/date-picker/tsconfig.json
@@ -10,6 +10,6 @@
     "strict": true,
     "target": "ESNext"
   },
-  "include": ["", "src", "../../toolkit/hooks/src/use-date-helper.tsx"],
+  "include": ["", "src", "../../toolkit/hooks/src/use-date-helpers.tsx"],
   "exclude": ["dist/**", "src/main.tsx"]
 }
