+++ development/date-picker-app/pages/date-picker/index.mdx
@@ -65,6 +65,7 @@ The `HeadlessDatePicker` component accepts the following props:
 - `calendarHeaderButtonsWrapper` : CSS classes for the header buttons wrapper.
 - `headerButtonClasses` : CSS classes for the header buttons.
 - `singleCalenderSectionWrapper` : CSS classes for the single calendar section wrapper.
+- `calendarWeeksClasses` : CSS classes for the calendar weeks section .
 - `monthNameClasses` : CSS classes for the month name.
 - `weekNameWrapper` : CSS classes for the week name wrapper.
 - `weekNameClasses` : CSS classes for the week name.
@@ -115,6 +116,7 @@ You can style the component using Tailwind CSS and CSS. Feel free to customize t
 | `singleCalenderSectionWrapper` | `string` | -            | CSS classes for the single calendar section wrapper. |
 | `monthNameClasses`          | `string`   | -             | CSS classes for the month name.                       |
 | `weekNameWrapper`           | `string`   | -             | CSS classes for the week name wrapper.                |
+| `calendarWeeksClasses`      | `string`   | -             | CSS classes for the calendar weeks wrapper.                |
 | `weekNameClasses`           | `string`   | -             | CSS classes for the week name.                        |
 | `selectableClasses`         | `string`   | -             | CSS classes for selectable dates.                     |
 | `selectedOrTodayClasses`    | `string`   | -             | CSS classes for selected or today's dates.            |
