+++ development/toolkits-app/pages/hooks/useClosableComponent.mdx
