+++ toolkit/hooks/index.ts
@@ -7,6 +7,6 @@ export { useProgress } from "./src/use-progress";
 export { useSlider } from "./src/use-slider";
 export { useBetaForm } from "./src/use-beta-form/index";
 export { useDataTable } from "./src/use-data-table";
-export { useDateHelpers } from "./src/use-date-helper";
+export { useDateHelpers } from "./src/use-date-helpers";
 export { default as useClosableComponent } from "./src/use-closable-component";
 export { default as useMountComponent } from "./src/use-mount-component";
